# perf: use db.count() instead of len(get_all)