Before:

```
%timeit -r 100 -n 100 get_open_count("My DocType", "My Docname")
2.14 ms ± 406 µs per loop (mean ± std. dev. of 100 runs, 100 loops each)
```

After:

```
%timeit -r 100 -n 100 get_open_count("My DocType", "My Docname")
1.72 ms ± 512 µs per loop (mean ± std. dev. of 100 runs, 100 loops each)
```

_My DocType_ has ~800 records, 20 of which are linked to _My Docname_.

### Thoughts

This function is leaking information (not introduced by this PR) since users will see the total number of records, not just of the ones they are permitted to see. Maybe a `frappe.get_list(dt, fields="COUNT(*)", filters=filters)` would be safer (but slower) here.