@barredterra 2023-07-26
Closing because I can't figure out how to implement this properly.